refactor(coeus-cuda): Split math dispatch leaves
