+++ CHECKLIST.md
@@ -8,6 +8,8 @@
       backend-parity workflow.
 - [x] Repair the stale CUDA/WGPU fallible pool wrapper return contracts that
       block the exact-head provider matrix.
+- [x] Split the mixed CUDA math dispatcher into elementwise, matmul, and
+      reduction leaves without changing the backend method surface.
 - [ ] Record exact-head WGPU, CUDA, ROCm, and Metal CI evidence without making
       an unmeasured runtime or resident-memory claim.
 
@@ -21,8 +23,8 @@ Owner: Codex on `codex/coeus-cuda-common-activation-parity`; claimed scope:
 `crates/coeus-wgpu/src/backend/ops/{pool,impls/pool}.rs`, plus the migrated
 WGPU pooling contract tests under
 `crates/coeus-wgpu/tests/wgpu_ops/backend/wgpu/pooling.rs`. The live peer owns
-the dirty fused-dispatch, reduction, backend-error, lockfile, ADR, and
-changelog files; they remain outside this increment.
+the dirty reduction, backend-error, and lockfile files; they remain outside
+this increment.
 
 Implementation head `a8dcc51c` is complete. The pool repair propagates WGPU
 kernel errors, removes the CUDA 2-D/3-D host fallback module, and returns typed
